Add ability to tweak OPTIONS reply with a custom http.Handler
We have a requirement that our front end application validates the origin
for all requests including the preflight request for OPTIONS.

Currently the httprouter package can respond to the OPTIONS requests with
the correct 'allows' data. However, it doesn't provide the ability to modify the origin,
accept types, or add custom headers.

This change adds the ability to to modify the request prior to responding with
the OPTIONS reply that httprouter already handles.

Useful links with detailed information regarding the above:
https://developer.mozilla.org/en-US/docs/Web/HTTP/Access_control_CORS
https://github.com/rs/cors