Params via Go 1.7 Contexts (#147)
* Supports Params in net/http Handlers with go1.7+

Go 1.7 introduces request-scoped context, providing a place to hang
Params even in net/http standard Handlers attached with Router.Handler
or Router.HandlerFunc.

This provides an alternative implementation of Router.Handler which
attaches the params to the request context. It also includes a bit of
new public API: ParamsKey is the context key used, and
ParamsFromContext uses ParamsKey to get Params from a context.Context.

All of this is guarded by a "go1.7" build tag. With go 1.6 and older
none of the new API is available and Params continue to be thrown away
in Handler/HandlerFunc-attached endpoints.

* WithValue is the 'context' function

* use an un-exported struct{} alias to collision-proof the key

* no longer a const

* update README.md to reflect additional allocations with the stdlib-compatible api